Correct bitmask corresponding to all threads active in a warp(64)

+++ src/gromacs/ewald/pme_gather_hip.cpp
@@ -95,7 +95,7 @@ __device__ __forceinline__ void reduce_atom_forces(float3* __restrict__ sm_force
 {
     if (!(order & (order - 1))) // Only for orders of power of 2
     {
-        const unsigned int activeMask = c_fullWarpMask;
+        const unsigned long activeMask = c_fullWarpMask;
 
         // A tricky shuffle reduction inspired by reduce_force_j_warp_shfl
         // TODO: find out if this is the best in terms of transactions count
